Move test to save cli ### What's done: * Move test to save cli
