+++ save-core/build.gradle.kts
@@ -3,7 +3,6 @@ import org.cqfn.save.generation.configFilePath
 import org.cqfn.save.generation.generateConfigOptions
 
 import de.undercouch.gradle.tasks.download.Download
-import org.gradle.nativeplatform.platform.internal.DefaultNativePlatform.getCurrentOperatingSystem
 import org.jetbrains.kotlin.gradle.dsl.KotlinCompile
 
 plugins {
@@ -19,7 +18,6 @@ kotlin {
                 implementation(project(":save-reporters"))
                 api("com.squareup.okio:okio-multiplatform:${Versions.okio}")
                 implementation("org.jetbrains.kotlinx:kotlinx-serialization-core:${Versions.Kotlinx.serialization}")
-                implementation("org.jetbrains.kotlinx:kotlinx-serialization-json:${Versions.Kotlinx.serialization}")
                 implementation("org.jetbrains.kotlinx:kotlinx-cli:${Versions.Kotlinx.cli}")
                 implementation("com.akuleshov7:ktoml-core:${Versions.ktoml}")
                 implementation("com.akuleshov7:ktoml-file:${Versions.ktoml}")
@@ -79,17 +77,6 @@ tasks.withType<KotlinCompile<*>>().forEach {
     it.dependsOn(generateVersionFileTaskProvider)
 }
 
-// Integration test should be able to have access to binary during the execution
-val os = getCurrentOperatingSystem()
-tasks.getByName("jvmTest").dependsOn(
-    when {
-        os.isLinux -> ":save-cli:linkDebugExecutableLinuxX64"
-        os.isWindows -> ":save-cli:linkDebugExecutableMingwX64"
-        os.isMacOsX -> ":save-cli:linkDebugExecutableMacosX64"
-        else -> throw GradleException("Unknown operating system $os")
-    }
-)
-
 tasks.register<Download>("downloadTestResources") {
     src(listOf(
         Versions.IntegrationTest.ktlintLink,
@@ -103,6 +90,7 @@ tasks.register<Download>("downloadTestResources") {
     }
 }
 val cleanupTask = tasks.register("cleanupTestResources") {
+    this.dependsOn(":save-cli:jvmTest")
     mustRunAfter(tasks.withType<Test>())
     doFirst {
         file("../examples/kotlin-diktat/ktlint").delete()
